# docs/swagger.yaml
basePath: /
definitions:
...
  request.UpdateAbsenRequest:
    properties:
      absen:
        type: integer
      absen_flag:
        type: integer
      lokasi:
        type: string
    type: object
...
  response.Response:
    properties:
      data: {}
      message:
        type: string
      status:
        type: string
    type: object
info:
  contact: {}
  description: API yang digunakan untuk website SIPEKOM
  title: API Sipekom
  version: 0.0.1
paths:
...
  /api/absen/{id_user}:
    get:
      description: get all Absen, mahasiswa have limited access
      parameters:
      - description: ID User
        in: path
        name: id_user
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: "get all Absen [mahasiswa \U0001F9F1, guest\U0001F512]."
      tags:
      - Absen
  /api/absen/create/{uri_base64}:
    get:
      description: get Absen by location.
      parameters:
      - description: location base64
        in: path
        name: location_base64
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: "create Absen [guest\U0001F512]."
      tags:
      - Absen
  /api/absen/delete/{id_absen}:
    delete:
      description: delete Absen by ID, only admin can delete absen.
      parameters:
      - description: ID Absen
        in: path
        name: id_absen
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: "delete Absen [konsulen\U0001F512, mahasiswa\U0001F512, guest\U0001F512]."
      tags:
      - Absen
  /api/absen/get/{id_absen}:
    get:
      description: get Absen by ID Absen, mahasiswa have limited access.
      parameters:
      - description: ID Absen
        in: path
        name: id_absen
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: "get Absen [mahasiswa \U0001F9F1, guest\U0001F512]."
      tags:
      - Absen
  /api/absen/update/{id_absen}:
    put:
      description: update Absen by ID, only Admin can update Absen.
      parameters:
      - description: body
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/request.UpdateAbsenRequest'
      - description: ID Absen
        in: path
        name: id_absen
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: "update Absen [konsulen\U0001F512, mahasiswa\U0001F512, guest\U0001F512]."
      tags:
      - Absen
  /api/data/{search_query}:
    get:
      consumes:
      - '*/*'
      description: get data of user.
      parameters:
      - description: Search Query
        in: path
        name: search_query
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      summary: "get user data \U0001F9F1."
      tags:
      - API
...
  /api/qr/get/{id_lokasi}:
    get:
      consumes:
      - '*/*'
      description: get qr codes based on id_lokasi.
      parameters:
      - description: ID Lokasi
        in: path
        name: id_lokasi
        required: true
        type: integer
      responses:
        "200":
          description: OK
      security:
      - ApiKeyAuth: []
      summary: "qr code image [mahasiswa\U0001F512, konsulen\U0001F512, guest\U0001F512]."
      tags:
      - API
...
  /api/whoami:
    get:
      consumes:
      - '*/*'
      description: get validation of the token.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/response.Response'
      security:
      - ApiKeyAuth: []
      summary: "check token validation [guest\U0001F512]."
      tags:
      - API
securityDefinitions:
  ApiKeyAuth:
    description: API Token
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"